+++ src/Model/Entities/ItemPedido.java
@@ -1,16 +1,18 @@
 package Model.Entities;
 
+import java.math.BigDecimal;
+
 public class ItemPedido {
     private int idPedido;
     private int idProduto;
     private int quantidade;
-    private double preco_unitario;
-    private double subtotal;
+    private BigDecimal preco_unitario;
+    private BigDecimal subtotal;
 
     public ItemPedido() {}
 
     // Sem ID separado — a PK aqui é composta (idPedido + idProduto), não há AUTO_INCREMENT
-    public ItemPedido(int idPedido, int idProduto, int quantidade, double preco_unitario, double subtotal) {
+    public ItemPedido(int idPedido, int idProduto, int quantidade, BigDecimal preco_unitario, BigDecimal subtotal) {
         this.idPedido = idPedido;
         this.idProduto = idProduto;
         this.quantidade = quantidade;
@@ -21,11 +23,11 @@ public class ItemPedido {
     public int getIdPedido() {return idPedido;}
     public int getIdProduto() {return idProduto;}
     public int getQuantidade() {return quantidade;}
-    public double getPreco_unitario() {return preco_unitario;}
-    public double getSubtotal() {return subtotal;}
+    public BigDecimal getPreco_unitario() {return preco_unitario;}
+    public BigDecimal getSubtotal() {return subtotal;}
     public void setQuantidade(int quantidade) {this.quantidade = quantidade;}
-    public void setPreco_unitario(double preco_unitario) {this.preco_unitario = preco_unitario;}
-    public void setSubtotal(double subtotal) {this.subtotal = subtotal;}
+    public void setPreco_unitario(BigDecimal preco_unitario) {this.preco_unitario = preco_unitario;}
+    public void setSubtotal(BigDecimal subtotal) {this.subtotal = subtotal;}
     public void setIdPedido(int idPedido) {this.idPedido = idPedido;}
     public void setIdProduto(int idProduto) {this.idProduto = idProduto;}
