+++ src/Model/Entities/Pedido.java
@@ -1,18 +1,19 @@
 package Model.Entities;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;;
 
 public class Pedido {
     private int idPedido;
     private int idCliente;
     private LocalDate data;    //(AAAA-MM-DD)
-    private double total;
+    private BigDecimal total;
     private String status;
 
     public Pedido() {};
 
     // Construtor completo para busca no Banco.
-    public Pedido(int idPedido, int idCliente, LocalDate data, double total, String status) {
+    public Pedido(int idPedido, int idCliente, LocalDate data, BigDecimal total, String status) {
         this.idPedido = idPedido;
         this.idCliente = idCliente;
         this.data = data;
@@ -20,7 +21,7 @@ public class Pedido {
         this.status = status;
     }
     //Construtor para Insert no banco(não uso o Id pq já é auto-incrementado pelo Banco)
-    public Pedido(LocalDate data, double total, String status) {
+    public Pedido(LocalDate data, BigDecimal total, String status) {
         this.data = data;
         this.total = total;
         this.status = status;
@@ -29,11 +30,11 @@ public class Pedido {
     public int getIdPedido() {return idPedido;}
     public int getIdCliente() {return idCliente;}
     public LocalDate getData() {return data;}
-    public double getTotal() {return total;}
+    public BigDecimal getTotal() {return total;}
     public String getStatus() {return status;}
     public void setStatus(String status) {this.status = status;}
     public void setData(LocalDate data) {this.data = data;}
-    public void setTotal(double total) {this.total = total;}
+    public void setTotal(BigDecimal total) {this.total = total;}
     public void setIdCliente(int idCliente) {this.idCliente = idCliente;}
     public void setIdPedido(int idPedido) {this.idPedido = idPedido;}
